+++ crates/aeronet_websocket/src/session/mod.rs
@@ -172,11 +172,7 @@ pub(crate) fn poll(mut sessions: Query<(Entity, &mut Session, &mut WebSocketIo)>
             session.recv.push(packet);
         }
 
-        trace!(
-            num_packets = num_packets.0,
-            num_bytes = num_bytes.0,
-            "Received packets",
-        );
+        trace!(%num_packets, %num_bytes, "Received packets");
     }
 }
 
@@ -200,10 +196,6 @@ fn flush(mut sessions: Query<(Entity, &mut Session, &WebSocketIo)>) {
             _ = io.send_packet_f2b.unbounded_send(packet);
         }
 
-        trace!(
-            num_packets = num_packets.0,
-            num_bytes = num_bytes.0,
-            "Flushed packets",
-        );
+        trace!(%num_packets, %num_bytes, "Flushed packets");
     }
 }